+++ src/mrgbstats/atoms.clj
@@ -2,5 +2,4 @@
 
 ; ----------
 ; Warning-related atoms
- (def warning-given-no-parallel-permtest-multiprops-atom (atom false))